+++ test/integration/server.rb
@@ -45,7 +45,7 @@ rs.resource_groups.create(
   location: 'eastus'
 )
 
-storage.storage_accounts.create(
+storage_account = storage.storage_accounts.create(
   name: 'fogstorageac',
   location: 'eastus',
   resource_group: 'TestRG-VM',
@@ -116,7 +116,14 @@ virtual_machine.detach_data_disk('datadisk1')
 ######################                                Delete Data Disk                            ######################
 ########################################################################################################################
 
-storage.delete_disk('TestRG-VM', 'fogstorageac', 'datadisk1')
+access_key = storage_account.get_access_keys[0].value
+Fog::Logger.debug access_key.inspect
+storage_data = Fog::Storage.new(
+    provider: 'AzureRM',
+    azure_storage_account_name: storage_account.name,
+    azure_storage_access_key: access_key
+)
+storage_data.delete_disk('datadisk1')
 
 ########################################################################################################################
 ######################                      List VM in a resource group                           ######################